[18] extract getMyProfile
